FIX: multimedia: utils: Fix search for video device
When a video file was passed as parameter to run a demo, the demo would
search for a video capture device, if it couldn't find one then the demo
would be terminated, even though a valid video file was used.

This patch fixes that issue, no longer searching for a video capture
device when a valid video file is used.

Reported-by: Jinghui Liu <jinghui.liu@nxp.com>
Signed-off-by: Alifer Moraes <alifer.moraes@nxp.com>